feat/PRSD-369: Review mark-ups refactor tests
