+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordRegistrationJourneyTests.kt
@@ -2,24 +2,20 @@ package uk.gov.communities.prsdb.webapp.integration
 
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import uk.gov.communities.prsdb.webapp.constants.enums.JourneyType
-import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.PhoneNumberFormPage
 
 class LandlordRegistrationJourneyTests : IntegrationTest() {
-    val journeyType = JourneyType.LANDLORD_REGISTRATION
-
     @Nested
     inner class LandlordRegistrationStepEmail {
         @Test
         fun `Submitting a valid email address redirects to the next step`() {
-            val formPage = navigator.goToEmailFormPage(journeyType)
+            val formPage = navigator.goToLandlordRegistrationEmailFormPage()
             formPage.fillEmail("test@example.com")
-            formPage.submit<PhoneNumberFormPage>()
+            formPage.submit()
         }
 
         @Test
         fun `Submitting an empty e-mail address returns an error`() {
-            val formPage = navigator.goToEmailFormPage(journeyType)
+            val formPage = navigator.goToLandlordRegistrationEmailFormPage()
             formPage.fillEmail("")
             formPage.submitUnsuccessfully()
             formPage.assertEmailFormErrorContains("Enter a valid email address to continue. An email is required for contact purposes.")
@@ -27,7 +23,7 @@ class LandlordRegistrationJourneyTests : IntegrationTest() {
 
         @Test
         fun `Submitting an invalid e-mail address returns an error`() {
-            val formPage = navigator.goToEmailFormPage(journeyType)
+            val formPage = navigator.goToLandlordRegistrationEmailFormPage()
             formPage.fillEmail("notAnEmail")
             formPage.submitUnsuccessfully()
             formPage.assertEmailFormErrorContains("Enter an email address in the right format")
