+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageobjects/pages/landlordRegistrationJourneyPages/PhoneNumberFormPageLandlordRegistration.kt
@@ -1,9 +1,10 @@
-package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages
+package uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.landlordRegistrationJourneyPages
 
 import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
+import uk.gov.communities.prsdb.webapp.integration.pageobjects.pages.BasePage
 
-class PhoneNumberFormPage(
+class PhoneNumberFormPageLandlordRegistration(
     page: Page,
 ) : BasePage(page) {
     override fun validate() {
